Move options to their own files
